+++ libraries/AP_Compass/AP_Compass.cpp
