+++ libraries/AP_Compass/AP_Compass_Backend.cpp
@@ -1,7 +1,8 @@
 /// -*- tab-width: 4; Mode: C++; c-basic-offset: 4; indent-tabs-mode: nil -*-
 
 #include <AP_HAL/AP_HAL.h>
-#include "Compass.h"
+
+#include "AP_Compass.h"
 #include "AP_Compass_Backend.h"
 
 extern const AP_HAL::HAL& hal;
