+++ libraries/AP_Compass/AP_Compass_Calibration.cpp
@@ -1,8 +1,9 @@
 /// -*- tab-width: 4; Mode: C++; c-basic-offset: 4; indent-tabs-mode: nil -*-
 #include <AP_HAL/AP_HAL.h>
-#include "Compass.h"
 #include <AP_Notify/AP_Notify.h>
 #include <GCS_MAVLink/GCS.h>
+
+#include "AP_Compass.h"
 
 extern AP_HAL::HAL& hal;
