+++ libraries/AP_Compass/AP_Compass_HMC5843.h
@@ -6,7 +6,7 @@
 #include <AP_HAL/I2CDevice.h>
 #include <AP_Math/AP_Math.h>
 
-#include "Compass.h"
+#include "AP_Compass.h"
 #include "AP_Compass_Backend.h"
 
 class AuxiliaryBus;
