+++ libraries/AP_HAL_SITL/SITL_State.h
@@ -17,7 +17,7 @@
 
 #include <AP_Baro/AP_Baro.h>
 #include <AP_InertialSensor/AP_InertialSensor.h>
-#include <AP_Compass/Compass.h>
+#include <AP_Compass/AP_Compass.h>
 #include <AP_OpticalFlow/AP_OpticalFlow.h>
 #include <AP_Terrain/AP_Terrain.h>
 #include <SITL/SITL.h>
